Move BufferedStreamReader to Common
