+++ src/Common/Common.csproj
@@ -17,6 +17,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="System.Text.Json" Version="6.0.6" />
+	  <PackageReference Include="Microsoft.IO.RecyclableMemoryStream" Version="2.2.1" />
 	</ItemGroup>
 
 </Project>
